[code cleanup] move packages specific functions to impl/packages.go
